chore: Move `std::regex` instances into `.cc` files
